Move auth dependency out of routes
